Move end to end test data to repo. (#214)
Ticket: #205

Put end to end test data under version control. DB files can be very large so we are using `git lfs` to store them as pointers.

The reason for this change is that tests should be updated in sync alongside the code. The prior method of storing all test data globally on S3 is a mess that will not scale well at all if more than one person is making changes.  We originally used S3 to avoid storing large DB files in the repo, but `git lfs` solves this problem.